backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.0
sqlalchemy[asyncio]==2.0.21
asyncpg==0.28.0
alembic==1.10.4
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.24.1